Sync changes from EE excluding enterprise directory
