+++ app/client/src/ce/constants/ReduxActionConstants.tsx
@@ -452,6 +452,8 @@ const ActionTypes = {
   EXECUTE_PLUGIN_ACTION_REQUEST: "EXECUTE_PLUGIN_ACTION_REQUEST",
   EXECUTE_PLUGIN_ACTION_SUCCESS: "EXECUTE_PLUGIN_ACTION_SUCCESS",
   SET_ACTION_RESPONSE_DISPLAY_FORMAT: "SET_ACTION_RESPONSE_DISPLAY_FORMAT",
+  SET_CURL_MODAL_OPEN: "SET_CURL_MODAL_OPEN",
+  SET_CURL_MODAL_CLOSE: "SET_CURL_MODAL_CLOSE",
   SUBMIT_CURL_FORM_INIT: "SUBMIT_CURL_FORM_INIT",
   SUBMIT_CURL_FORM_SUCCESS: "SUBMIT_CURL_FORM_SUCCESS",
   SET_WIDGET_DYNAMIC_PROPERTY: "SET_WIDGET_DYNAMIC_PROPERTY",
