+++ app/client/src/ce/pages/Editor/CurlImport/index.ts
@@ -1,0 +1,1 @@
+export { default as AppCURLImportModal } from "./AppCURLImportModal";
