+++ app/client/src/ee/pages/Editor/CurlImport/index.ts
@@ -1,0 +1,3 @@
+import { AppCURLImportModal as CE_AppCURLImportModal } from "ce/pages/Editor/CurlImport";
+
+export { CE_AppCURLImportModal as AppCURLImportModal };
