--- app/client/src/pages/Editor/CurlImport/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./CurlImportEditor";
