+++ app/client/src/pages/Editor/IntegrationEditor/CreateNewDatasourceTab.tsx
@@ -23,7 +23,6 @@ import MockDataSources from "./MockDataSources";
 import NewApiScreen from "./NewApi";
 import NewQueryScreen from "./NewQuery";
 import { isAirgapped } from "@appsmith/utils/airgapHelpers";
-import history from "utils/history";
 import { showDebuggerFlag } from "selectors/debuggerSelectors";
 import {
   createMessage,
@@ -85,7 +84,6 @@ function UseMockDatasources({ active, mockDatasources }: MockDataSourcesProps) {
 
 function CreateNewAPI({
   active,
-  history,
   isCreating,
   isOnboardingScreen,
   pageId,
@@ -111,7 +109,6 @@ function CreateNewAPI({
     <div id="new-api" ref={newAPIRef}>
       <Text kind="heading-m">APIs</Text>
       <NewApiScreen
-        history={history}
         isCreating={isCreating}
         isOnboardingScreen={isOnboardingScreen}
         location={location}
@@ -125,7 +122,6 @@ function CreateNewAPI({
 
 function CreateNewDatasource({
   active,
-  history,
   isCreating,
   isOnboardingScreen,
   pageId,
@@ -159,7 +155,6 @@ function CreateNewDatasource({
       <NewQueryScreen
         editorId={editorId}
         editorType={editorType}
-        history={history}
         isAirgappedInstance={isAirgappedInstance}
         isCreating={isCreating}
         location={location}
@@ -174,7 +169,6 @@ function CreateNewDatasource({
 
 function CreateNewSaasIntegration({
   active,
-  history,
   isCreating,
   pageId,
   showUnsupportedPluginDialog,
@@ -202,7 +196,6 @@ function CreateNewSaasIntegration({
       <div id="new-saas-api" ref={newSaasAPIRef}>
         <Text kind="heading-m">SaaS integrations</Text>
         <NewApiScreen
-          history={history}
           isCreating={isCreating}
           location={location}
           pageId={pageId}
@@ -215,7 +208,6 @@ function CreateNewSaasIntegration({
 }
 
 function CreateNewAIIntegration({
-  history,
   isCreating,
   pageId,
   showUnsupportedPluginDialog,
@@ -228,7 +220,6 @@ function CreateNewAIIntegration({
       <div id="new-ai-query">
         <Text kind="heading-m">AI integrations</Text>
         <AIDataSources
-          history={history}
           isCreating={isCreating}
           location={location}
           pageId={pageId}
@@ -304,7 +295,6 @@ class CreateNewDatasourceTab extends React.Component<
             )}
           <CreateNewDatasource
             active={false}
-            history={history}
             isCreating={isCreating}
             isOnboardingScreen={!!isOnboardingScreen}
             location={location}
@@ -315,7 +305,6 @@ class CreateNewDatasourceTab extends React.Component<
           <StyledDivider />
           <CreateNewAPI
             active={false}
-            history={history}
             isCreating={isCreating}
             isOnboardingScreen={!!isOnboardingScreen}
             location={location}
@@ -325,7 +314,6 @@ class CreateNewDatasourceTab extends React.Component<
           <StyledDivider />
           <CreateNewDatasource
             active={false}
-            history={history}
             isCreating={isCreating}
             location={location}
             pageId={pageId}
@@ -333,14 +321,12 @@ class CreateNewDatasourceTab extends React.Component<
           />
           <CreateNewSaasIntegration
             active={false}
-            history={history}
             isCreating={isCreating}
             location={location}
             pageId={pageId}
             showUnsupportedPluginDialog={this.showUnsupportedPluginDialog}
           />
           <CreateNewAIIntegration
-            history={history}
             isCreating={isCreating}
             pageId={pageId}
             showUnsupportedPluginDialog={this.showUnsupportedPluginDialog}
