+++ app/client/src/pages/Editor/IntegrationEditor/NewApi.tsx
@@ -6,16 +6,13 @@ import {
   createTempDatasourceFromForm,
 } from "actions/datasourceActions";
 import type { AppState } from "@appsmith/reducers";
-import CurlLogo from "assets/images/Curl-logo.svg";
 import PlusLogo from "assets/images/Plus-logo.svg";
 import type { GenerateCRUDEnabledPluginMap, Plugin } from "api/PluginApi";
 import AnalyticsUtil from "@appsmith/utils/AnalyticsUtil";
-import { CURL } from "constants/AppsmithActionConstants/ActionConstants";
 import { PluginPackageName, PluginType } from "entities/Action";
 import { getQueryParams } from "utils/URLUtils";
 import { getGenerateCRUDEnabledPluginMap } from "@appsmith/selectors/entitiesSelector";
 import { getIsGeneratePageInitiator } from "utils/GenerateCrudUtil";
-import { curlImportPageURL } from "@appsmith/RouteBuilder";
 import { getAssetUrl } from "@appsmith/utils/airgapHelpers";
 import { Spinner } from "design-system";
 import { useEditorType } from "@appsmith/hooks";
@@ -120,10 +117,6 @@ export const CardContentWrapper = styled.div`
 `;
 
 interface ApiHomeScreenProps {
-  history: {
-    replace: (data: string) => void;
-    push: (data: string) => void;
-  };
   location: {
     search: string;
   };
@@ -155,14 +148,8 @@ export const API_ACTION = {
 };
 
 function NewApiScreen(props: Props) {
-  const {
-    history,
-    isCreating,
-    isOnboardingScreen,
-    pageId,
-    plugins,
-    showSaasAPIs,
-  } = props;
+  const { isCreating, isOnboardingScreen, pageId, plugins, showSaasAPIs } =
+    props;
   const editorType = useEditorType(location.pathname);
   const { editorId, parentEntityId, parentEntityType } =
     useParentEntityInfo(editorType);
@@ -230,26 +217,6 @@ function NewApiScreen(props: Props) {
       case API_ACTION.CREATE_NEW_GRAPHQL_API:
         handleCreateNew(actionType);
         break;
-      case API_ACTION.IMPORT_CURL: {
-        AnalyticsUtil.logEvent("IMPORT_API_CLICK", {
-          importSource: CURL,
-        });
-        AnalyticsUtil.logEvent("CREATE_DATA_SOURCE_CLICK", {
-          source: CURL,
-        });
-
-        delete queryParams.isGeneratePageMode;
-        const curlImportURL = curlImportPageURL({
-          pageId,
-          params: {
-            from: "datasources",
-            ...queryParams,
-          },
-        });
-
-        history.push(curlImportURL);
-        break;
-      }
       case API_ACTION.CREATE_DATASOURCE_FORM: {
         props.createTempDatasourceFromForm({
           pluginId: params.pluginId,
@@ -347,21 +314,6 @@ function NewApiScreen(props: Props) {
             </CardContentWrapper>
           </ApiCard>
         ))}
-        {!showSaasAPIs && (
-          <ApiCard
-            className="t--createBlankCurlCard"
-            onClick={() => handleOnClick(API_ACTION.IMPORT_CURL)}
-          >
-            <CardContentWrapper>
-              <img
-                alt="CURL"
-                className="curlImage t--curlImage content-icon"
-                src={CurlLogo}
-              />
-              <p className="textBtn">CURL import</p>
-            </CardContentWrapper>
-          </ApiCard>
-        )}
       </ApiCardsContainer>
     </StyledContainer>
   );
