+++ app/client/src/pages/Editor/index.tsx
@@ -47,7 +47,7 @@ import ReconfigureCDKeyModal from "@appsmith/components/gitComponents/Reconfigur
 import DisableCDModal from "@appsmith/components/gitComponents/DisableCDModal";
 import { PartialExportModal } from "components/editorComponents/PartialImportExport/PartialExportModal";
 import { PartialImportModal } from "components/editorComponents/PartialImportExport/PartialImportModal";
-import CurlImport from "pages/Editor/CurlImport";
+import { AppCURLImportModal } from "@appsmith/pages/Editor/CurlImport";
 
 interface EditorProps {
   currentApplicationId?: string;
@@ -184,7 +184,7 @@ class Editor extends Component<Props> {
             <SignpostingOverlay />
             <PartialExportModal />
             <PartialImportModal />
-            <CurlImport />
+            <AppCURLImportModal />
           </GlobalHotKeys>
         </div>
         <RequestConfirmationModal />
