+++ app/client/src/selectors/curlImportSelectors.ts
@@ -1,0 +1,7 @@
+import type { AppState } from "@appsmith/reducers";
+
+export const getIsImportingCurl = (state: AppState) =>
+  state.ui.imports.isImportingCurl;
+
+export const getIsCurlModalOpen = (state: AppState) =>
+  state.ui.imports.isCurlModalOpen;
